tried to suppress some compilation warnings in reactphysics3d. There are probably several more type casts to do, but I tried with what the compiler reported.

+++ include/reactphysics3d/collision/ConvexMesh.h
@@ -135,7 +135,7 @@ class ConvexMesh {
  * @return The number of vertices in the mesh
  */
 RP3D_FORCE_INLINE uint32 ConvexMesh::getNbVertices() const {
-    return mVertices.size();
+    return static_cast<uint32>(mVertices.size());
 }
 
 /// Return a vertex
